Make startup components more generic for reuse on touchscreen
